Lisätty Procfile
Procfile is a mechanism for declaring what commands are run by your
application’s dynos on the Heroku platform. It follows the process
model. You can use a Procfile to declare various process types, such as
multiple types of workers, a singleton process like a clock, or a
consumer of the Twitter streaming API.